moved storylines outside of stageView
